+++ api/src/main/java/org/allaymc/api/math/location/Location3d.java
@@ -24,122 +24,99 @@ public class Location3d extends Position3d implements Location3dc {
 
     public double pitch;
     public double yaw;
-    public double headYaw;
 
     public Location3d(Location3dc l) {
         super(l);
         this.pitch = l.pitch();
         this.yaw = l.yaw();
-        this.headYaw = l.headYaw();
     }
 
     public Location3d(Location3ic l) {
         super(l.x(), l.y(), l.z(), l.dimension());
         this.pitch = l.pitch();
         this.yaw = l.yaw();
-        this.headYaw = l.headYaw();
     }
 
     public Location3d(double x, double y, double z, Dimension dimension) {
         super(x, y, z, dimension);
     }
 
-    public Location3d(double d, double pitch, double yaw, double headYaw, Dimension dimension) {
-        super(d, dimension);
-        this.pitch = pitch;
-        this.yaw = yaw;
-        this.headYaw = headYaw;
-    }
-
-    public Location3d(double x, double y, double z, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(double x, double y, double z, double pitch, double yaw, Dimension dimension) {
         super(x, y, z, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(Vector3fc v, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(Vector3fc v, double pitch, double yaw, Dimension dimension) {
         super(v, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(Vector3ic v, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(Vector3ic v, double pitch, double yaw, Dimension dimension) {
         super(v, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(Vector2fc v, double z, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(Vector2fc v, double z, double pitch, double yaw, Dimension dimension) {
         super(v, z, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(Vector2ic v, double z, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(Vector2ic v, double z, double pitch, double yaw, Dimension dimension) {
         super(v, z, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(Vector3dc v, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(Vector3dc v, double pitch, double yaw, Dimension dimension) {
         super(v, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(Vector2dc v, double z, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(Vector2dc v, double z, double pitch, double yaw, Dimension dimension) {
         super(v, z, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(double[] xyz, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(double[] xyz, double pitch, double yaw, Dimension dimension) {
         super(xyz, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(float[] xyz, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(float[] xyz, double pitch, double yaw, Dimension dimension) {
         super(xyz, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(ByteBuffer buffer, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(ByteBuffer buffer, double pitch, double yaw, Dimension dimension) {
         super(buffer, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(int index, ByteBuffer buffer, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(int index, ByteBuffer buffer, double pitch, double yaw, Dimension dimension) {
         super(index, buffer, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(DoubleBuffer buffer, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(DoubleBuffer buffer, double pitch, double yaw, Dimension dimension) {
         super(buffer, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
-    public Location3d(int index, DoubleBuffer buffer, double pitch, double yaw, double headYaw, Dimension dimension) {
+    public Location3d(int index, DoubleBuffer buffer, double pitch, double yaw, Dimension dimension) {
         super(index, buffer, dimension);
         this.pitch = pitch;
         this.yaw = yaw;
-        this.headYaw = headYaw;
     }
 
     @Override
@@ -153,18 +130,12 @@ public class Location3d extends Position3d implements Location3dc {
     }
 
     @Override
-    public double headYaw() {
-        return headYaw;
-    }
-
-    @Override
     public void writeExternal(ObjectOutput out) throws IOException {
         out.writeDouble(x);
         out.writeDouble(y);
         out.writeDouble(z);
         out.writeDouble(pitch);
         out.writeDouble(yaw);
-        out.writeDouble(headYaw);
         out.writeUTF(this.dimension().getWorld().getWorldData().getDisplayName());
         out.writeInt(this.dimension().getDimensionInfo().dimensionId());
     }
@@ -176,7 +147,6 @@ public class Location3d extends Position3d implements Location3dc {
         z = in.readDouble();
         pitch = in.readDouble();
         yaw = in.readDouble();
-        headYaw = in.readDouble();
         dimension = new WeakReference<>(Server.getInstance().getWorldPool().getWorld(in.readUTF()).getDimension(in.readInt()));
     }
 
@@ -185,12 +155,12 @@ public class Location3d extends Position3d implements Location3dc {
         if (this == o) return true;
         if (!(o instanceof Location3dc that)) return false;
         if (!super.equals(o)) return false;
-        return Objects.equal(dimension.get(), that.dimension()) && Objects.equal(pitch, that.pitch()) && Objects.equal(yaw, that.yaw()) && Objects.equal(headYaw, that.headYaw());
+        return Objects.equal(dimension.get(), that.dimension()) && Objects.equal(pitch, that.pitch()) && Objects.equal(yaw, that.yaw());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(super.hashCode(), pitch, yaw, headYaw, dimension.get());
+        return Objects.hashCode(super.hashCode(), pitch, yaw, dimension.get());
     }
 
     @Override
@@ -200,6 +170,6 @@ public class Location3d extends Position3d implements Location3dc {
 
     @Override
     public String toString(NumberFormat formatter) {
-        return "(" + Runtime.format(x, formatter) + " " + Runtime.format(y, formatter) + " " + Runtime.format(z, formatter) + " pitch=" + this.pitch + " yaw=" + this.yaw + " headYaw=" + this.headYaw + " dimension=" + this.dimension().getWorld().getWorldData().getDisplayName() + this.dimension().getDimensionInfo().dimensionId() + ")";
+        return "(" + Runtime.format(x, formatter) + " " + Runtime.format(y, formatter) + " " + Runtime.format(z, formatter) + " pitch=" + this.pitch + " yaw=" + this.yaw + " dimension=" + this.dimension().getWorld().getWorldData().getDisplayName() + this.dimension().getDimensionInfo().dimensionId() + ")";
     }
 }