+++ api/src/main/java/org/allaymc/api/math/location/Location3fc.java
@@ -9,6 +9,4 @@ public interface Location3fc extends Position3fc {
     double pitch();
 
     double yaw();
-
-    double headYaw();
 }