+++ api/src/main/java/org/allaymc/api/math/location/Location3ic.java
@@ -10,5 +10,4 @@ public interface Location3ic extends Position3ic {
 
     double yaw();
 
-    double headYaw();
 }